refactor: use autogenerated service for /professionals endpoint
