+++ src/app/master/pages/master-tag-edit/master-tag-edit.module.ts
@@ -2,7 +2,6 @@ import { CommonModule } from '@angular/common';
 import { NgModule } from '@angular/core';
 import { FormsModule, ReactiveFormsModule } from '@angular/forms';
 import { TagsSelectInputComponent } from '@app/master/components/tags-select-input/tags-select-input.component';
-import { TagsListApiService } from '@app/master/services/tags-list-api.service';
 import { SharedModule } from '@app/shared/shared.module';
 import { IonicModule } from '@ionic/angular';
 import { TranslateModule } from '@ngx-translate/core';
@@ -20,7 +19,6 @@ import { MasterTagEditPage } from './master-tag-edit.page';
     TranslateModule,
   ],
   declarations: [MasterTagEditPage, TagsSelectInputComponent],
-  providers: [TagsListApiService],
 })
 export class MasterTagEditPageModule {
 }
