+++ src/app/master/services/master-photos-generator-factory.service.ts
@@ -1,9 +1,9 @@
 import { Injectable } from '@angular/core';
 import { MasterManagerService } from '@app/core/services/master-manager.service';
-import { MasterPhoto } from '@app/master/models/master-photo';
-import { MasterPhotosReadonlyApiService } from '@app/master/services/master-photos-readonly-api.service';
+import { ProfessionalsService } from '@app/api/services';
 import { Observable } from 'rxjs';
 import { map, switchMap } from 'rxjs/operators';
+import { ProfessionalPhotoList } from '@app/api/models';
 
 @Injectable({
   providedIn: 'root',
@@ -11,16 +11,16 @@ import { map, switchMap } from 'rxjs/operators';
 export class MasterPhotosGeneratorFactoryService {
 
   constructor(
-    private readonly masterPhotosReadonlyApi: MasterPhotosReadonlyApiService,
+    private readonly professionalsApi: ProfessionalsService,
     private readonly masterManager: MasterManagerService,
   ) {
   }
 
-  public getPhotos(masterId?: number): Observable<MasterPhoto[]> {
+  public getPhotos(masterId?: number): Observable<ProfessionalPhotoList[]> {
     return masterId ? this.get(masterId) : this.masterManager.getMasterList().pipe(switchMap(list => this.get(list[0].id)));
   }
 
-  private get(masterId: number): Observable<MasterPhoto[]> {
-    return this.masterPhotosReadonlyApi.get({ professional: masterId.toString() }).pipe(map(res => res.results));
+  private get(masterId: number): Observable<ProfessionalPhotoList[]> {
+    return this.professionalsApi.professionalsProfessionalPhotosList({ professional: masterId.toString() }).pipe(map(res => res.results));
   }
 }
