+++ src/environments/environment.prod.ts.dist
@@ -18,7 +18,6 @@ export const environment = {
         user_language: '/api/accounts/languages/',
         location: '/api/accounts/locations/',
         master: '/api/accounts/professionals/',
-        master_list: '/api/professionals/professionals/',
         plugin: '',
         reset_password_link: '/api/accounts/send-reset-password-link/',
         reset_password: '/api/accounts/reset-password/',
@@ -31,12 +30,9 @@ export const environment = {
         regions: '/api/location/regions/',
         subregions: '/api/location/subregions/',
         send_verify_registration: '/api/accounts/resend-verify-registration/',
-        subcategory: '/api/professionals/subcategories/',
-        category: '/api/professionals/categories/',
         verify_registration: '/api/accounts/verify-registration/',
         professional_contact: '/api/accounts/professional-contacts/',
         professional_tags: '/api/accounts/professional-tags/',
-        professional_tags_list: '/api/professionals/tags/',
         professional_location: '/api/accounts/professional-locations/',
         reviews: '/api/accounts/reviews/',
         reviews_readonly: '/api/communication/reviews',
@@ -59,7 +55,6 @@ export const environment = {
         master_schedule: '/api/accounts/professional-schedule/',
         services_readonly: '/api/services/services/',
         service_tag_readonly: '/api/services/tags/',
-        master_photos_readonly: '/api/professionals/professional-photos/',
         sent_orders: '/api/accounts/orders/sent/',
         received_orders: '/api/accounts/orders/received/',
         calendar: '/api/schedule/calendar/',
